+++ samples/nuxt3/components/menu.vue
@@ -1,0 +1,16 @@
+<template>
+    <nav>
+        <ul class="twic-sample-nav-links">
+            <li>
+                <a href="/">
+                    Static
+                </a>
+            </li>
+            <li>
+                <a href="/ssr">
+                    Ssr
+                </a>
+            </li>
+        </ul>
+    </nav>
+</template>