+++ samples/nuxt3/pages/index.vue
@@ -1,0 +1,14 @@
+<template>
+    <main>
+        <Menu />
+        <Sample />
+    </main>
+</template>
+<script>
+import Sample from "@twicpics/components-sample/Sample.vue";
+export default {
+    "components": {
+        Sample,
+    },
+};
+</script>