+++ src/nuxt2/plugin.js
@@ -1,5 +1,5 @@
 import Vue from "vue";
-import TwicPics from "@twicpics/components/vueXXX";
+import TwicPics from "@twicpics/components/vue2";
 const options = JSON.parse( `<%= JSON.stringify(options) %>` );
 
 export default () => {