+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # 0.6.7
 - __chores__: improves configuration of `transition` property
+- __feature__: supports `Nuxt.js 3.x`
 
 ## 0.6.6
 - __bug__: fixes a refresh issue in `angular`